+++ src/SmbSpeedDoctor.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace SmbSpeedDoctor.Core;
-
-public class Class1
-{
-
-}